Refactor lib.rs roughly by feature
Maintains existing API compatibility
