+++ src/lib.rs
@@ -21,29 +21,84 @@ able to parse any URI, such as `urn:isbn:0451450523`.
 extern crate bitflags;
 
 use serde::{Deserialize, Serialize};
-use serde_json;
 use serde_repr::{Deserialize_repr, Serialize_repr};
 
 pub use url::Url;
 
-use std::borrow::Cow;
-
-#[cfg(feature = "proposed")]
-use std::convert::TryFrom;
-
 use std::collections::HashMap;
 
-#[cfg(feature = "proposed")]
-use base64;
 use serde::de;
 use serde::de::Error as Error_;
 use serde_json::Value;
 
-#[cfg(feature = "proposed")]
-use serde::ser::SerializeSeq;
-
 pub mod notification;
 pub mod request;
+
+#[cfg(feature = "proposed")]
+mod call_hierarchy;
+#[cfg(feature = "proposed")]
+pub use call_hierarchy::*;
+
+mod code_action;
+pub use code_action::*;
+
+mod code_lens;
+pub use code_lens::*;
+
+mod color;
+pub use color::*;
+
+mod completion;
+pub use completion::*;
+
+mod document_highlight;
+pub use document_highlight::*;
+
+mod document_link;
+pub use document_link::*;
+
+mod document_symbols;
+pub use document_symbols::*;
+
+mod folding_range;
+pub use folding_range::*;
+
+mod formatting;
+pub use formatting::*;
+
+mod hover;
+pub use hover::*;
+
+mod progress;
+pub use progress::*;
+
+mod references;
+pub use references::*;
+
+mod rename;
+pub use rename::*;
+
+pub mod selection_range;
+pub use selection_range::*;
+
+#[cfg(feature = "proposed")]
+mod semantic_highlighting;
+#[cfg(feature = "proposed")]
+pub use semantic_highlighting::*;
+
+#[cfg(feature = "proposed")]
+mod semantic_tokens;
+#[cfg(feature = "proposed")]
+pub use semantic_tokens::*;
+
+mod signature_help;
+pub use signature_help::*;
+
+mod workspace_folders;
+pub use workspace_folders::*;
+
+mod workspace_symbols;
+pub use workspace_symbols::*;
 
 /* ----------------- Auxiliary types ----------------- */
 
@@ -302,44 +357,6 @@ pub struct TextDocumentEdit {
 
     /// The edits to be applied.
     pub edits: Vec<TextEdit>,
-}
-
-/// A special text edit to provide an insert and a replace operation.
-///
-/// @since 3.16.0 - Proposed state
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct InsertReplaceEdit {
-    /// The string to be inserted.
-    pub new_text: String,
-
-    /// The range if the insert is requested
-    pub insert: Range,
-
-    /// The range if the replace is requested.
-    pub replace: Range,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(untagged)]
-pub enum CompletionTextEdit {
-    Edit(TextEdit),
-    #[cfg(feature = "proposed")]
-    InsertAndReplace(InsertReplaceEdit),
-}
-
-impl From<TextEdit> for CompletionTextEdit {
-    fn from(edit: TextEdit) -> Self {
-        CompletionTextEdit::Edit(edit)
-    }
-}
-
-#[cfg(feature = "proposed")]
-impl From<InsertReplaceEdit> for CompletionTextEdit {
-    fn from(edit: InsertReplaceEdit) -> Self {
-        CompletionTextEdit::InsertAndReplace(edit)
-    }
 }
 
 /// Options to create a file.
@@ -853,58 +870,6 @@ pub struct WorkspaceEditCapability {
     pub failure_handling: Option<FailureHandlingKind>,
 }
 
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkspaceCapability {
-    /// The server supports workspace folder.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub workspace_folders: Option<WorkspaceFolderCapability>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkspaceFolderCapability {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub supported: Option<bool>,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub change_notifications: Option<WorkspaceFolderCapabilityChangeNotifications>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(untagged)]
-pub enum WorkspaceFolderCapabilityChangeNotifications {
-    Bool(bool),
-    Id(String),
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkspaceFolder {
-    /// The associated URI for this workspace folder.
-    pub uri: Url,
-    /// The name of the workspace folder. Defaults to the uri's basename.
-    pub name: String,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DidChangeWorkspaceFoldersParams {
-    /// The actual workspace folder change event.
-    pub event: WorkspaceFoldersChangeEvent,
-}
-
-/// The workspace folder change event.
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkspaceFoldersChangeEvent {
-    /// The array of added workspace folders
-    pub added: Vec<WorkspaceFolder>,
-
-    /// The array of the removed workspace folders
-    pub removed: Vec<WorkspaceFolder>,
-}
-
 #[derive(Debug, Eq, PartialEq, Deserialize, Serialize, Copy, Clone)]
 #[serde(rename_all = "lowercase")]
 pub enum ResourceOperationKind {
@@ -922,6 +887,41 @@ pub enum FailureHandlingKind {
     Undo,
 }
 
+/// A symbol kind.
+#[derive(Debug, Eq, PartialEq, Copy, Clone, Serialize_repr, Deserialize_repr)]
+#[repr(u8)]
+pub enum SymbolKind {
+    File = 1,
+    Module = 2,
+    Namespace = 3,
+    Package = 4,
+    Class = 5,
+    Method = 6,
+    Property = 7,
+    Field = 8,
+    Constructor = 9,
+    Enum = 10,
+    Interface = 11,
+    Function = 12,
+    Variable = 13,
+    Constant = 14,
+    String = 15,
+    Number = 16,
+    Boolean = 17,
+    Array = 18,
+    Object = 19,
+    Key = 20,
+    Null = 21,
+    EnumMember = 22,
+    Struct = 23,
+    Event = 24,
+    Operator = 25,
+    TypeParameter = 26,
+
+    // Capturing all unknown enums by this lib.
+    Unknown = 255,
+}
+
 /// Specific capabilities for the `SymbolKind` in the `workspace/symbol` request.
 #[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
@@ -935,31 +935,6 @@ pub struct SymbolKindCapability {
     /// the symbol kinds from `File` to `Array` as defined in
     /// the initial version of the protocol.
     pub value_set: Option<Vec<SymbolKind>>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkspaceSymbolClientCapabilities {
-    /// This capability supports dynamic registration.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub dynamic_registration: Option<bool>,
-
-    /// Specific capabilities for the `SymbolKind` in the `workspace/symbol` request.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub symbol_kind: Option<SymbolKindCapability>,
-
-    /// The client supports tags on `SymbolInformation`.
-    /// Clients supporting tags have to handle unknown tags gracefully.
-    ///
-    /// @since 3.16.0
-    ///
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        deserialize_with = "TagSupport::deserialize_compat"
-    )]
-    #[cfg(feature = "proposed")]
-    pub tag_support: Option<TagSupport<SymbolTag>>,
 }
 
 /// Workspace specific client capabilities.
@@ -1020,151 +995,6 @@ pub struct SynchronizationCapability {
     /// The client supports did save notifications.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub did_save: Option<bool>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CompletionItemCapability {
-    /// Client supports snippets as insert text.
-    ///
-    /// A snippet can define tab stops and placeholders with `$1`, `$2`
-    /// and `${3:foo}`. `$0` defines the final tab stop, it defaults to
-    /// the end of the snippet. Placeholders with equal identifiers are linked,
-    /// that is typing in one will update others too.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub snippet_support: Option<bool>,
-
-    /// Client supports commit characters on a completion item.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub commit_characters_support: Option<bool>,
-
-    /// Client supports the follow content formats for the documentation
-    /// property. The order describes the preferred format of the client.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub documentation_format: Option<Vec<MarkupKind>>,
-
-    /// Client supports the deprecated property on a completion item.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub deprecated_support: Option<bool>,
-
-    /// Client supports the preselect property on a completion item.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub preselect_support: Option<bool>,
-
-    /// Client supports the tag property on a completion item. Clients supporting
-    /// tags have to handle unknown tags gracefully. Clients especially need to
-    /// preserve unknown tags when sending a completion item back to the server in
-    /// a resolve call.
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        deserialize_with = "TagSupport::deserialize_compat"
-    )]
-    pub tag_support: Option<TagSupport<CompletionItemTag>>,
-
-    /// Client support insert replace edit to control different behavior if a
-    /// completion item is inserted in the text or should replace text.
-    ///
-    /// @since 3.16.0 - Proposed state
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[cfg(feature = "proposed")]
-    pub insert_replace_support: Option<bool>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize_repr, Serialize_repr)]
-#[repr(u8)]
-pub enum CompletionItemTag {
-    Deprecated = 1,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CompletionItemKindCapability {
-    /// The completion item kind values the client supports. When this
-    /// property exists the client also guarantees that it will
-    /// handle values outside its set gracefully and falls back
-    /// to a default value when unknown.
-    ///
-    /// If this property is not present the client only supports
-    /// the completion items kinds from `Text` to `Reference` as defined in
-    /// the initial version of the protocol.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub value_set: Option<Vec<CompletionItemKind>>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct HoverCapability {
-    /// Whether completion supports dynamic registration.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub dynamic_registration: Option<bool>,
-
-    /// Client supports the follow content formats for the content
-    /// property. The order describes the preferred format of the client.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub content_format: Option<Vec<MarkupKind>>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CompletionCapability {
-    /// Whether completion supports dynamic registration.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub dynamic_registration: Option<bool>,
-
-    /// The client supports the following `CompletionItem` specific
-    /// capabilities.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub completion_item: Option<CompletionItemCapability>,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub completion_item_kind: Option<CompletionItemKindCapability>,
-
-    /// The client supports to send additional context information for a
-    /// `textDocument/completion` requestion.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub context_support: Option<bool>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SignatureInformationSettings {
-    /// Client supports the follow content formats for the documentation
-    /// property. The order describes the preferred format of the client.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub documentation_format: Option<Vec<MarkupKind>>,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub parameter_information: Option<ParameterInformationSettings>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ParameterInformationSettings {
-    /// The client supports processing label offsets instead of a
-    /// simple label string.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub label_offset_support: Option<bool>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SignatureHelpCapability {
-    /// Whether completion supports dynamic registration.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub dynamic_registration: Option<bool>,
-
-    /// The client supports the following `SignatureInformation`
-    /// specific properties.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub signature_information: Option<SignatureInformationSettings>,
-
-    /// The client supports to send additional context information for a
-    /// `textDocument/signatureHelp` request. A client that opts into
-    /// contextSupport will also support the `retriggerCharacters` on
-    /// `SignatureHelpOptions`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub context_support: Option<bool>,
 }
 
 #[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
@@ -1385,135 +1215,6 @@ pub enum TextDocumentSyncKind {
     Incremental = 2,
 }
 
-/// Completion options.
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CompletionOptions {
-    /// The server provides support to resolve additional information for a completion item.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub resolve_provider: Option<bool>,
-
-    /// The characters that trigger completion automatically.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub trigger_characters: Option<Vec<String>>,
-
-    #[serde(flatten)]
-    pub work_done_progress_options: WorkDoneProgressOptions,
-}
-
-/// Hover options.
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct HoverOptions {
-    #[serde(flatten)]
-    pub work_done_progress_options: WorkDoneProgressOptions,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct HoverRegistrationOptions {
-    #[serde(flatten)]
-    pub text_document_registration_options: TextDocumentRegistrationOptions,
-
-    #[serde(flatten)]
-    pub hover_options: HoverOptions,
-}
-
-/// Signature help options.
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SignatureHelpOptions {
-    /// The characters that trigger signature help automatically.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub trigger_characters: Option<Vec<String>>,
-
-    ///  List of characters that re-trigger signature help.
-    /// These trigger characters are only active when signature help is already showing. All trigger characters
-    /// are also counted as re-trigger characters.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub retrigger_characters: Option<Vec<String>>,
-
-    #[serde(flatten)]
-    pub work_done_progress_options: WorkDoneProgressOptions,
-}
-
-/// Signature help options.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-pub struct SignatureHelpRegistrationOptions {
-    #[serde(flatten)]
-    pub text_document_registration_options: TextDocumentRegistrationOptions,
-}
-/// Signature help options.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize_repr, Serialize_repr)]
-#[repr(u8)]
-pub enum SignatureHelpTriggerKind {
-    /// Signature help was invoked manually by the user or by a command.
-    Invoked = 1,
-    ///  Signature help was triggered by a trigger character.
-    TriggerCharacter = 2,
-    /// Signature help was triggered by the cursor moving or by the document content changing.
-    ContentChange = 3,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SignatureHelpParams {
-    /// The signature help context. This is only available if the client specifies
-    /// to send this using the client capability  `textDocument.signatureHelp.contextSupport === true`
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub context: Option<SignatureHelpContext>,
-
-    #[serde(flatten)]
-    pub text_document_position_params: TextDocumentPositionParams,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SignatureHelpContext {
-    ///  Action that caused signature help to be triggered.
-    pub trigger_kind: SignatureHelpTriggerKind,
-
-    /// Character that caused signature help to be triggered.
-    /// This is undefined when `triggerKind !== SignatureHelpTriggerKind.TriggerCharacter`
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub trigger_character: Option<String>,
-
-    /// `true` if signature help was already showing when it was triggered.
-    /// Retriggers occur when the signature help is already active and can be caused by actions such as
-    /// typing a trigger character, a cursor move, or document content changes.
-    pub is_retrigger: bool,
-
-    /// The currently active `SignatureHelp`.
-    /// The `activeSignatureHelp` has its `SignatureHelp.activeSignature` field updated based on
-    /// the user navigating through available signatures.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub active_signature_help: Option<SignatureHelp>,
-}
-
-/// Code Lens options.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodeLensOptions {
-    /// Code lens has a resolve provider as well.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub resolve_provider: Option<bool>,
-}
-
-/// Format document on type options
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentOnTypeFormattingOptions {
-    /// A character on which formatting should be triggered, like `}`.
-    pub first_trigger_character: String,
-
-    /// More trigger characters.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub more_trigger_character: Option<Vec<String>>,
-}
-
 /// Execute command options.
 #[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
 pub struct ExecuteCommandOptions {
@@ -1632,106 +1333,6 @@ impl From<bool> for TypeDefinitionProviderCapability {
     fn from(from: bool) -> Self {
         Self::Simple(from)
     }
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(untagged)]
-pub enum HoverProviderCapability {
-    Simple(bool),
-    Options(HoverOptions),
-}
-
-impl From<HoverOptions> for HoverProviderCapability {
-    fn from(from: HoverOptions) -> Self {
-        Self::Options(from)
-    }
-}
-
-impl From<bool> for HoverProviderCapability {
-    fn from(from: bool) -> Self {
-        Self::Simple(from)
-    }
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(untagged)]
-pub enum ColorProviderCapability {
-    Simple(bool),
-    ColorProvider(ColorProviderOptions),
-    Options(StaticTextDocumentColorProviderOptions),
-}
-
-impl From<ColorProviderOptions> for ColorProviderCapability {
-    fn from(from: ColorProviderOptions) -> Self {
-        Self::ColorProvider(from)
-    }
-}
-
-impl From<StaticTextDocumentColorProviderOptions> for ColorProviderCapability {
-    fn from(from: StaticTextDocumentColorProviderOptions) -> Self {
-        Self::Options(from)
-    }
-}
-
-impl From<bool> for ColorProviderCapability {
-    fn from(from: bool) -> Self {
-        Self::Simple(from)
-    }
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(untagged)]
-pub enum CodeActionProviderCapability {
-    Simple(bool),
-    Options(CodeActionOptions),
-}
-
-impl From<CodeActionOptions> for CodeActionProviderCapability {
-    fn from(from: CodeActionOptions) -> Self {
-        Self::Options(from)
-    }
-}
-
-impl From<bool> for CodeActionProviderCapability {
-    fn from(from: bool) -> Self {
-        Self::Simple(from)
-    }
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodeActionCapability {
-    ///
-    /// This capability supports dynamic registration.
-    ///
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub dynamic_registration: Option<bool>,
-
-    /// The client support code action literals as a valid
-    /// response of the `textDocument/codeAction` request.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub code_action_literal_support: Option<CodeActionLiteralSupport>,
-
-    /// Whether code action supports the `isPreferred` property.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub is_preferred_support: Option<bool>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodeActionLiteralSupport {
-    /// The code action kind is support with the following value set.
-    pub code_action_kind: CodeActionKindLiteralSupport,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodeActionKindLiteralSupport {
-    /// The code action kind values the client supports. When this
-    /// property exists the client also guarantees that it will
-    /// handle values outside its set gracefully and falls back
-    /// to a default value when unknown.
-    pub value_set: Vec<String>,
 }
 
 #[derive(Debug, PartialEq, Clone, Default, Deserialize, Serialize)]
@@ -1854,18 +1455,6 @@ pub struct ServerCapabilities {
 }
 
 #[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentLinkCapabilities {
-    /// Whether document link supports dynamic registration.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub dynamic_registration: Option<bool>,
-
-    /// Whether the client support the `tooltip` property on `DocumentLink`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tooltip_support: Option<bool>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
 pub struct ShowMessageParams {
     /// The message type. See {@link MessageType}.
     #[serde(rename = "type")]
@@ -1959,21 +1548,6 @@ pub struct StaticRegistrationOptions {
 #[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct StaticTextDocumentRegistrationOptions {
-    /// A document selector to identify the scope of the registration. If set to null
-    /// the document selector provided on the client side will be used.
-    pub document_selector: Option<DocumentSelector>,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ColorProviderOptions {}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct StaticTextDocumentColorProviderOptions {
     /// A document selector to identify the scope of the registration. If set to null
     /// the document selector provided on the client side will be used.
     pub document_selector: Option<DocumentSelector>,
@@ -2221,262 +1795,11 @@ impl PublishDiagnosticsParams {
     }
 }
 
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub struct CompletionRegistrationOptions {
-    #[serde(flatten)]
-    pub text_document_registration_options: TextDocumentRegistrationOptions,
-
-    #[serde(flatten)]
-    pub completion_options: CompletionOptions,
-}
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum CompletionResponse {
-    Array(Vec<CompletionItem>),
-    List(CompletionList),
-}
-
-impl From<Vec<CompletionItem>> for CompletionResponse {
-    fn from(items: Vec<CompletionItem>) -> Self {
-        CompletionResponse::Array(items)
-    }
-}
-
-impl From<CompletionList> for CompletionResponse {
-    fn from(list: CompletionList) -> Self {
-        CompletionResponse::List(list)
-    }
-}
-
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CompletionParams {
-    // This field was "mixed-in" from TextDocumentPositionParams
-    #[serde(flatten)]
-    pub text_document_position: TextDocumentPositionParams,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-
-    // CompletionParams properties:
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub context: Option<CompletionContext>,
-}
-
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CompletionContext {
-    /// How the completion was triggered.
-    pub trigger_kind: CompletionTriggerKind,
-
-    /// The trigger character (a single character) that has trigger code complete.
-    /// Is undefined if `triggerKind !== CompletionTriggerKind.TriggerCharacter`
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub trigger_character: Option<String>,
-}
-
-/// How a completion was triggered.
-#[derive(Debug, PartialEq, Clone, Copy, Deserialize_repr, Serialize_repr)]
-#[repr(u8)]
-pub enum CompletionTriggerKind {
-    Invoked = 1,
-    TriggerCharacter = 2,
-    TriggerForIncompleteCompletions = 3,
-}
-
-/// Represents a collection of [completion items](#CompletionItem) to be presented
-/// in the editor.
-#[derive(Debug, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CompletionList {
-    /// This list it not complete. Further typing should result in recomputing
-    /// this list.
-    pub is_incomplete: bool,
-
-    /// The completion items.
-    pub items: Vec<CompletionItem>,
-}
-
 #[derive(Debug, Eq, PartialEq, Deserialize, Serialize, Clone)]
 #[serde(untagged)]
 pub enum Documentation {
     String(String),
     MarkupContent(MarkupContent),
-}
-
-#[derive(Debug, PartialEq, Default, Deserialize, Serialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct CompletionItem {
-    /// The label of this completion item. By default
-    /// also the text that is inserted when selecting
-    /// this completion.
-    pub label: String,
-
-    /// The kind of this completion item. Based of the kind
-    /// an icon is chosen by the editor.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub kind: Option<CompletionItemKind>,
-
-    /// A human-readable string with additional information
-    /// about this item, like type or symbol information.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub detail: Option<String>,
-
-    /// A human-readable string that represents a doc-comment.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub documentation: Option<Documentation>,
-
-    /// Indicates if this item is deprecated.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub deprecated: Option<bool>,
-
-    /// Select this item when showing.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub preselect: Option<bool>,
-
-    /// A string that shoud be used when comparing this item
-    /// with other items. When `falsy` the label is used.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sort_text: Option<String>,
-
-    /// A string that should be used when filtering a set of
-    /// completion items. When `falsy` the label is used.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub filter_text: Option<String>,
-
-    /// A string that should be inserted a document when selecting
-    /// this completion. When `falsy` the label is used.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub insert_text: Option<String>,
-
-    /// The format of the insert text. The format applies to both the `insertText` property
-    /// and the `newText` property of a provided `textEdit`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub insert_text_format: Option<InsertTextFormat>,
-
-    /// An edit which is applied to a document when selecting
-    /// this completion. When an edit is provided the value of
-    /// insertText is ignored.
-    ///
-    /// Most editors support two different operation when accepting a completion item. One is to insert a
-    /// completion text and the other is to replace an existing text with a competion text. Since this can
-    /// usually not predetermend by a server it can report both ranges. Clients need to signal support for
-    /// `InsertReplaceEdits` via the `textDocument.completion.insertReplaceSupport` client capability
-    /// property.
-    ///
-    /// *Note 1:* The text edit's range as well as both ranges from a insert replace edit must be a
-    /// [single line] and they must contain the position at which completion has been requested.
-    /// *Note 2:* If an `InsertReplaceEdit` is returned the edit's insert range must be a prefix of
-    /// the edit's replace range, that means it must be contained and starting at the same position.
-    ///
-    /// @since 3.16.0 additional type `InsertReplaceEdit` - Proposed state
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub text_edit: Option<CompletionTextEdit>,
-
-    /// An optional array of additional text edits that are applied when
-    /// selecting this completion. Edits must not overlap with the main edit
-    /// nor with themselves.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub additional_text_edits: Option<Vec<TextEdit>>,
-
-    /// An optional command that is executed *after* inserting this completion. *Note* that
-    /// additional modifications to the current document should be described with the
-    /// additionalTextEdits-property.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub command: Option<Command>,
-
-    /// An data entry field that is preserved on a completion item between
-    /// a completion and a completion resolve request.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub data: Option<Value>,
-
-    /// Tags for this completion item.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<Vec<CompletionItemTag>>,
-}
-
-impl CompletionItem {
-    /// Create a CompletionItem with the minimum possible info (label and detail).
-    pub fn new_simple(label: String, detail: String) -> CompletionItem {
-        CompletionItem {
-            label,
-            detail: Some(detail),
-            ..Self::default()
-        }
-    }
-}
-
-/// The kind of a completion entry.
-#[derive(Debug, Eq, PartialEq, Clone, Copy, Serialize_repr, Deserialize_repr)]
-#[repr(u8)]
-pub enum CompletionItemKind {
-    Text = 1,
-    Method = 2,
-    Function = 3,
-    Constructor = 4,
-    Field = 5,
-    Variable = 6,
-    Class = 7,
-    Interface = 8,
-    Module = 9,
-    Property = 10,
-    Unit = 11,
-    Value = 12,
-    Enum = 13,
-    Keyword = 14,
-    Snippet = 15,
-    Color = 16,
-    File = 17,
-    Reference = 18,
-    Folder = 19,
-    EnumMember = 20,
-    Constant = 21,
-    Struct = 22,
-    Event = 23,
-    Operator = 24,
-    TypeParameter = 25,
-}
-
-/// Defines how to interpret the insert text in a completion item
-#[derive(Debug, Eq, PartialEq, Clone, Copy, Serialize_repr, Deserialize_repr)]
-#[repr(u8)]
-pub enum InsertTextFormat {
-    PlainText = 1,
-    Snippet = 2,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct HoverParams {
-    #[serde(flatten)]
-    pub text_document_position_params: TextDocumentPositionParams,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-}
-
-/// The result of a hover request.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-pub struct Hover {
-    /// The hover's content
-    pub contents: HoverContents,
-    /// An optional range is a range inside a text document
-    /// that is used to visualize a hover, e.g. by changing the background color.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub range: Option<Range>,
-}
-
-/// Hover contents could be single entry or multiple entries.
-#[derive(Debug, Eq, PartialEq, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum HoverContents {
-    Scalar(MarkedString),
-    Array(Vec<MarkedString>),
-    Markup(MarkupContent),
 }
 
 /// The marked string is rendered:
@@ -2511,67 +1834,6 @@ impl MarkedString {
             value: code_block,
         })
     }
-}
-
-/// Signature help represents the signature of something
-/// callable. There can be multiple signature but only one
-/// active and only one active parameter.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SignatureHelp {
-    /// One or more signatures.
-    pub signatures: Vec<SignatureInformation>,
-
-    /// The active signature.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub active_signature: Option<i64>,
-
-    /// The active parameter of the active signature.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub active_parameter: Option<i64>,
-}
-
-/// Represents the signature of something callable. A signature
-/// can have a label, like a function-name, a doc-comment, and
-/// a set of parameters.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-pub struct SignatureInformation {
-    /// The label of this signature. Will be shown in
-    /// the UI.
-    pub label: String,
-
-    /// The human-readable doc-comment of this signature. Will be shown
-    /// in the UI but can be omitted.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub documentation: Option<Documentation>,
-
-    /// The parameters of this signature.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub parameters: Option<Vec<ParameterInformation>>,
-}
-
-/// Represents a parameter of a callable-signature. A parameter can
-/// have a label and a doc-comment.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-pub struct ParameterInformation {
-    /// The label of this parameter information.
-    ///
-    /// Either a string or an inclusive start and exclusive end offsets within its containing
-    /// signature label. (see SignatureInformation.label). *Note*: A label of type string must be
-    /// a substring of its containing signature label.
-    pub label: ParameterLabel,
-
-    /// The human-readable doc-comment of this parameter. Will be shown
-    /// in the UI but can be omitted.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub documentation: Option<Documentation>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(untagged)]
-pub enum ParameterLabel {
-    Simple(String),
-    LabelOffsets([u64; 2]),
 }
 
 #[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
@@ -2614,245 +1876,6 @@ impl From<Vec<LocationLink>> for GotoDefinitionResponse {
     }
 }
 
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ReferenceParams {
-    // Text Document and Position fields
-    #[serde(flatten)]
-    pub text_document_position: TextDocumentPositionParams,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-
-    // ReferenceParams properties:
-    pub context: ReferenceContext,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Copy, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ReferenceContext {
-    /// Include the declaration of the current symbol.
-    pub include_declaration: bool,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentHighlightParams {
-    #[serde(flatten)]
-    pub text_document_position_params: TextDocumentPositionParams,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-}
-
-/// A document highlight is a range inside a text document which deserves
-/// special attention. Usually a document highlight is visualized by changing
-/// the background color of its range.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-pub struct DocumentHighlight {
-    /// The range this highlight applies to.
-    pub range: Range,
-
-    /// The highlight kind, default is DocumentHighlightKind.Text.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub kind: Option<DocumentHighlightKind>,
-}
-
-/// A document highlight kind.
-#[derive(Debug, Eq, PartialEq, Copy, Clone, Deserialize_repr, Serialize_repr)]
-#[repr(u8)]
-pub enum DocumentHighlightKind {
-    /// A textual occurrance.
-    Text = 1,
-
-    /// Read-access of a symbol, like reading a variable.
-    Read = 2,
-
-    /// Write-access of a symbol, like writing to a variable.
-    Write = 3,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentSymbolClientCapabilities {
-    /// This capability supports dynamic registration.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub dynamic_registration: Option<bool>,
-
-    /// Specific capabilities for the `SymbolKind`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub symbol_kind: Option<SymbolKindCapability>,
-
-    /// The client support hierarchical document symbols.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub hierarchical_document_symbol_support: Option<bool>,
-
-    /// The client supports tags on `SymbolInformation`. Tags are supported on
-    /// `DocumentSymbol` if `hierarchicalDocumentSymbolSupport` is set to true.
-    /// Clients supporting tags have to handle unknown tags gracefully.
-    ///
-    /// @since 3.16.0
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        deserialize_with = "TagSupport::deserialize_compat"
-    )]
-    #[cfg(feature = "proposed")]
-    pub tag_support: Option<TagSupport<SymbolTag>>,
-}
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum DocumentSymbolResponse {
-    Flat(Vec<SymbolInformation>),
-    Nested(Vec<DocumentSymbol>),
-}
-
-impl From<Vec<SymbolInformation>> for DocumentSymbolResponse {
-    fn from(info: Vec<SymbolInformation>) -> Self {
-        DocumentSymbolResponse::Flat(info)
-    }
-}
-
-impl From<Vec<DocumentSymbol>> for DocumentSymbolResponse {
-    fn from(symbols: Vec<DocumentSymbol>) -> Self {
-        DocumentSymbolResponse::Nested(symbols)
-    }
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentSymbolParams {
-    /// The text document.
-    pub text_document: TextDocumentIdentifier,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-}
-
-/// Represents programming constructs like variables, classes, interfaces etc.
-/// that appear in a document. Document symbols can be hierarchical and they have two ranges:
-/// one that encloses its definition and one that points to its most interesting range,
-/// e.g. the range of an identifier.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentSymbol {
-    /// The name of this symbol.
-    pub name: String,
-    /// More detail for this symbol, e.g the signature of a function. If not provided the
-    /// name is used.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub detail: Option<String>,
-    /// The kind of this symbol.
-    pub kind: SymbolKind,
-    /// Tags for this completion item.
-    ///  since 3.16.0
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[cfg(feature = "proposed")]
-    pub tags: Option<Vec<SymbolTag>>,
-    /// Indicates if this symbol is deprecated.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[deprecated(note = "Use tags instead")]
-    pub deprecated: Option<bool>,
-    /// The range enclosing this symbol not including leading/trailing whitespace but everything else
-    /// like comments. This information is typically used to determine if the the clients cursor is
-    /// inside the symbol to reveal in the symbol in the UI.
-    pub range: Range,
-    /// The range that should be selected and revealed when this symbol is being picked, e.g the name of a function.
-    /// Must be contained by the the `range`.
-    pub selection_range: Range,
-    /// Children of this symbol, e.g. properties of a class.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub children: Option<Vec<DocumentSymbol>>,
-}
-
-/// Represents information about programming constructs like variables, classes,
-/// interfaces etc.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SymbolInformation {
-    /// The name of this symbol.
-    pub name: String,
-
-    /// The kind of this symbol.
-    pub kind: SymbolKind,
-
-    /// Tags for this completion item.
-    ///  since 3.16.0
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[cfg(feature = "proposed")]
-    pub tags: Option<Vec<SymbolTag>>,
-
-    /// Indicates if this symbol is deprecated.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[deprecated(note = "Use tags instead")]
-    pub deprecated: Option<bool>,
-
-    /// The location of this symbol.
-    pub location: Location,
-
-    /// The name of the symbol containing this symbol.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub container_name: Option<String>,
-}
-
-/// A symbol kind.
-#[derive(Debug, Eq, PartialEq, Copy, Clone, Serialize_repr, Deserialize_repr)]
-#[repr(u8)]
-pub enum SymbolKind {
-    File = 1,
-    Module = 2,
-    Namespace = 3,
-    Package = 4,
-    Class = 5,
-    Method = 6,
-    Property = 7,
-    Field = 8,
-    Constructor = 9,
-    Enum = 10,
-    Interface = 11,
-    Function = 12,
-    Variable = 13,
-    Constant = 14,
-    String = 15,
-    Number = 16,
-    Boolean = 17,
-    Array = 18,
-    Object = 19,
-    Key = 20,
-    Null = 21,
-    EnumMember = 22,
-    Struct = 23,
-    Event = 24,
-    Operator = 25,
-    TypeParameter = 26,
-
-    // Capturing all unknown enums by this lib.
-    Unknown = 255,
-}
-
-/// The parameters of a Workspace Symbol Request.
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-pub struct WorkspaceSymbolParams {
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    /// A non-empty query string
-    pub query: String,
-}
-
 #[derive(Debug, PartialEq, Clone, Default, Deserialize, Serialize)]
 pub struct ExecuteCommandParams {
     /// The identifier of the actual command handler.
@@ -2885,668 +1908,6 @@ pub struct ApplyWorkspaceEditParams {
 pub struct ApplyWorkspaceEditResponse {
     /// Indicates whether the edit was applied or not.
     pub applied: bool,
-}
-
-/// Params for the CodeActionRequest
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodeActionParams {
-    /// The document in which the command was invoked.
-    pub text_document: TextDocumentIdentifier,
-
-    /// The range for which the command was invoked.
-    pub range: Range,
-
-    /// Context carrying additional information.
-    pub context: CodeActionContext,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-}
-
-/// response for CodeActionRequest
-pub type CodeActionResponse = Vec<CodeActionOrCommand>;
-
-#[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
-#[serde(untagged)]
-pub enum CodeActionOrCommand {
-    Command(Command),
-    CodeAction(CodeAction),
-}
-
-impl From<Command> for CodeActionOrCommand {
-    fn from(comand: Command) -> Self {
-        CodeActionOrCommand::Command(comand)
-    }
-}
-
-impl From<CodeAction> for CodeActionOrCommand {
-    fn from(action: CodeAction) -> Self {
-        CodeActionOrCommand::CodeAction(action)
-    }
-}
-
-#[derive(Debug, Eq, PartialEq, Hash, PartialOrd, Clone, Deserialize, Serialize)]
-pub struct CodeActionKind(Cow<'static, str>);
-
-impl CodeActionKind {
-    /// Empty kind.
-    pub const EMPTY: CodeActionKind = CodeActionKind::new("");
-
-    /// Base kind for quickfix actions: 'quickfix'
-    pub const QUICKFIX: CodeActionKind = CodeActionKind::new("quickfix");
-
-    /// Base kind for refactoring actions: 'refactor'
-    pub const REFACTOR: CodeActionKind = CodeActionKind::new("refactor");
-
-    /// Base kind for refactoring extraction actions: 'refactor.extract'
-    ///
-    /// Example extract actions:
-    ///
-    /// - Extract method
-    /// - Extract function
-    /// - Extract variable
-    /// - Extract interface from class
-    /// - ...
-    pub const REFACTOR_EXTRACT: CodeActionKind = CodeActionKind::new("refactor.extract");
-
-    /// Base kind for refactoring inline actions: 'refactor.inline'
-    ///
-    /// Example inline actions:
-    ///
-    /// - Inline function
-    /// - Inline variable
-    /// - Inline constant
-    /// - ...
-    pub const REFACTOR_INLINE: CodeActionKind = CodeActionKind::new("refactor.inline");
-
-    /// Base kind for refactoring rewrite actions: 'refactor.rewrite'
-    ///
-    /// Example rewrite actions:
-    ///
-    /// - Convert JavaScript function to class
-    /// - Add or remove parameter
-    /// - Encapsulate field
-    /// - Make method static
-    /// - Move method to base class
-    /// - ...
-    pub const REFACTOR_REWRITE: CodeActionKind = CodeActionKind::new("refactor.rewrite");
-
-    /// Base kind for source actions: `source`
-    ///
-    /// Source code actions apply to the entire file.
-    pub const SOURCE: CodeActionKind = CodeActionKind::new("source");
-
-    /// Base kind for an organize imports source action: `source.organizeImports`
-    pub const SOURCE_ORGANIZE_IMPORTS: CodeActionKind =
-        CodeActionKind::new("source.organizeImports");
-
-    pub const fn new(tag: &'static str) -> Self {
-        CodeActionKind(Cow::Borrowed(tag))
-    }
-
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl From<String> for CodeActionKind {
-    fn from(from: String) -> Self {
-        CodeActionKind(Cow::from(from))
-    }
-}
-
-impl From<&'static str> for CodeActionKind {
-    fn from(from: &'static str) -> Self {
-        CodeActionKind::new(from)
-    }
-}
-
-#[derive(Debug, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodeAction {
-    /// A short, human-readable, title for this code action.
-    pub title: String,
-
-    /// The kind of the code action.
-    /// Used to filter code actions.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub kind: Option<CodeActionKind>,
-
-    /// The diagnostics that this code action resolves.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub diagnostics: Option<Vec<Diagnostic>>,
-
-    /// The workspace edit this code action performs.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub edit: Option<WorkspaceEdit>,
-
-    /// A command this code action executes. If a code action
-    /// provides an edit and a command, first the edit is
-    /// executed and then the command.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub command: Option<Command>,
-
-    /// Marks this as a preferred action. Preferred actions are used by the `auto fix` command and can be targeted
-    /// by keybindings.
-    /// A quick fix should be marked preferred if it properly addresses the underlying error.
-    /// A refactoring should be marked preferred if it is the most reasonable choice of actions to take.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub is_preferred: Option<bool>,
-}
-
-/// Contains additional diagnostic information about the context in which
-/// a code action is run.
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-pub struct CodeActionContext {
-    /// An array of diagnostics.
-    pub diagnostics: Vec<Diagnostic>,
-
-    /// Requested kind of actions to return.
-    ///
-    /// Actions not of this kind are filtered out by the client before being shown. So servers
-    /// can omit computing them.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub only: Option<Vec<CodeActionKind>>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodeActionOptions {
-    /// CodeActionKinds that this server may return.
-    ///
-    /// The list of kinds may be generic, such as `CodeActionKind.Refactor`, or the server
-    /// may list out every specific kind they provide.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub code_action_kinds: Option<Vec<CodeActionKind>>,
-
-    #[serde(flatten)]
-    pub work_done_progress_options: WorkDoneProgressOptions,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodeLensParams {
-    /// The document to request code lens for.
-    pub text_document: TextDocumentIdentifier,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-}
-
-/// A code lens represents a command that should be shown along with
-/// source text, like the number of references, a way to run tests, etc.
-///
-/// A code lens is _unresolved_ when no command is associated to it. For performance
-/// reasons the creation of a code lens and resolving should be done in two stages.
-#[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
-pub struct CodeLens {
-    /// The range in which this code lens is valid. Should only span a single line.
-    pub range: Range,
-
-    /// The command this code lens represents.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub command: Option<Command>,
-
-    /// A data entry field that is preserved on a code lens item between
-    /// a code lens and a code lens resolve request.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub data: Option<Value>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentLinkParams {
-    /// The document to provide document links for.
-    pub text_document: TextDocumentIdentifier,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-}
-
-/// A document link is a range in a text document that links to an internal or external resource, like another
-/// text document or a web site.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-pub struct DocumentLink {
-    /// The range this link applies to.
-    pub range: Range,
-    /// The uri this link points to.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub target: Option<Url>,
-
-    /// The tooltip text when you hover over this link.
-    ///
-    /// If a tooltip is provided, is will be displayed in a string that includes instructions on how to
-    /// trigger the link, such as `{0} (ctrl + click)`. The specific instructions vary depending on OS,
-    /// user settings, and localization.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tooltip: Option<String>,
-
-    /// A data entry field that is preserved on a document link between a DocumentLinkRequest
-    /// and a DocumentLinkResolveRequest.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub data: Option<Value>,
-}
-
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentFormattingParams {
-    /// The document to format.
-    pub text_document: TextDocumentIdentifier,
-
-    /// The format options.
-    pub options: FormattingOptions,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-}
-
-/// Value-object describing what options formatting should use.
-#[derive(Debug, PartialEq, Clone, Default, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FormattingOptions {
-    /// Size of a tab in spaces.
-    pub tab_size: u64,
-
-    /// Prefer spaces over tabs.
-    pub insert_spaces: bool,
-
-    /// Signature for further properties.
-    #[serde(flatten)]
-    pub properties: HashMap<String, FormattingProperty>,
-
-    /// Trim trailing whitespaces on a line.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub trim_trailing_whitespace: Option<bool>,
-
-    /// Insert a newline character at the end of the file if one does not exist.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub insert_final_newline: Option<bool>,
-
-    /// Trim all newlines after the final newline at the end of the file.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub trim_final_newlines: Option<bool>,
-}
-
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(untagged)]
-pub enum FormattingProperty {
-    Bool(bool),
-    Number(f64),
-    String(String),
-}
-
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentRangeFormattingParams {
-    /// The document to format.
-    pub text_document: TextDocumentIdentifier,
-
-    /// The range to format
-    pub range: Range,
-
-    /// The format options
-    pub options: FormattingOptions,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-}
-
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentOnTypeFormattingParams {
-    /// Text Document and Position fields.
-    #[serde(flatten)]
-    pub text_document_position: TextDocumentPositionParams,
-
-    /// The character that has been typed.
-    pub ch: String,
-
-    /// The format options.
-    pub options: FormattingOptions,
-}
-
-/// Extends TextDocumentRegistrationOptions
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentOnTypeFormattingRegistrationOptions {
-    /// A document selector to identify the scope of the registration. If set to null
-    /// the document selector provided on the client side will be used.
-    pub document_selector: Option<DocumentSelector>,
-
-    /// A character on which formatting should be triggered, like `}`.
-    pub first_trigger_character: String,
-
-    /// More trigger characters.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub more_trigger_character: Option<Vec<String>>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RenameParams {
-    /// Text Document and Position fields
-    #[serde(flatten)]
-    pub text_document_position: TextDocumentPositionParams,
-
-    /// The new name of the symbol. If the given name is not valid the
-    /// request must return a [ResponseError](#ResponseError) with an
-    /// appropriate message set.
-    pub new_name: String,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(untagged)]
-pub enum RenameProviderCapability {
-    Simple(bool),
-    Options(RenameOptions),
-}
-
-impl From<RenameOptions> for RenameProviderCapability {
-    fn from(from: RenameOptions) -> Self {
-        Self::Options(from)
-    }
-}
-
-impl From<bool> for RenameProviderCapability {
-    fn from(from: bool) -> Self {
-        Self::Simple(from)
-    }
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RenameOptions {
-    /// Renames should be checked and tested before being executed.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub prepare_provider: Option<bool>,
-
-    #[serde(flatten)]
-    pub work_done_progress_options: WorkDoneProgressOptions,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RenameCapability {
-    /// Whether rename supports dynamic registration.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub dynamic_registration: Option<bool>,
-
-    /// Client supports testing for validity of rename operations before execution.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub prepare_support: Option<bool>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(untagged)]
-pub enum PrepareRenameResponse {
-    Range(Range),
-    RangeWithPlaceholder { range: Range, placeholder: String },
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentLinkOptions {
-    /// Document links have a resolve provider as well.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub resolve_provider: Option<bool>,
-
-    #[serde(flatten)]
-    pub work_done_progress_options: WorkDoneProgressOptions,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DocumentColorParams {
-    /// The text document
-    pub text_document: TextDocumentIdentifier,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-}
-
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ColorInformation {
-    /// The range in the document where this color appears.
-    pub range: Range,
-    /// The actual color value for this color range.
-    pub color: Color,
-}
-
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Color {
-    /// The red component of this color in the range [0-1].
-    pub red: f64,
-    /// The green component of this color in the range [0-1].
-    pub green: f64,
-    /// The blue component of this color in the range [0-1].
-    pub blue: f64,
-    /// The alpha component of this color in the range [0-1].
-    pub alpha: f64,
-}
-
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ColorPresentationParams {
-    /// The text document.
-    pub text_document: TextDocumentIdentifier,
-
-    /// The color information to request presentations for.
-    pub color: Color,
-
-    /// The range where the color would be inserted. Serves as a context.
-    pub range: Range,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-}
-
-#[derive(Debug, PartialEq, Eq, Deserialize, Serialize, Default, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct ColorPresentation {
-    /// The label of this color presentation. It will be shown on the color
-    /// picker header. By default this is also the text that is inserted when selecting
-    /// this color presentation.
-    pub label: String,
-
-    /// An [edit](#TextEdit) which is applied to a document when selecting
-    /// this presentation for the color.  When `falsy` the [label](#ColorPresentation.label)
-    /// is used.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub text_edit: Option<TextEdit>,
-
-    /// An optional array of additional [text edits](#TextEdit) that are applied when
-    /// selecting this color presentation. Edits must not overlap with the main [edit](#ColorPresentation.textEdit) nor with themselves.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub additional_text_edits: Option<Vec<TextEdit>>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FoldingRangeParams {
-    /// The text document.
-    pub text_document: TextDocumentIdentifier,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(untagged)]
-pub enum FoldingRangeProviderCapability {
-    Simple(bool),
-    FoldingProvider(FoldingProviderOptions),
-    Options(StaticTextDocumentColorProviderOptions),
-}
-
-impl From<StaticTextDocumentColorProviderOptions> for FoldingRangeProviderCapability {
-    fn from(from: StaticTextDocumentColorProviderOptions) -> Self {
-        Self::Options(from)
-    }
-}
-
-impl From<FoldingProviderOptions> for FoldingRangeProviderCapability {
-    fn from(from: FoldingProviderOptions) -> Self {
-        Self::FoldingProvider(from)
-    }
-}
-
-impl From<bool> for FoldingRangeProviderCapability {
-    fn from(from: bool) -> Self {
-        Self::Simple(from)
-    }
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-pub struct FoldingProviderOptions {}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FoldingRangeCapability {
-    /// Whether implementation supports dynamic registration for folding range providers. If this is set to `true`
-    /// the client supports the new `(FoldingRangeProviderOptions & TextDocumentRegistrationOptions & StaticRegistrationOptions)`
-    /// return value for the corresponding server capability as well.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub dynamic_registration: Option<bool>,
-
-    /// The maximum number of folding ranges that the client prefers to receive per document. The value serves as a
-    /// hint, servers are free to follow the limit.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub range_limit: Option<u64>,
-    /// If set, the client signals that it only supports folding complete lines. If set, client will
-    /// ignore specified `startCharacter` and `endCharacter` properties in a FoldingRange.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub line_folding_only: Option<bool>,
-}
-
-/// Represents a folding range.
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FoldingRange {
-    /// The zero-based line number from where the folded range starts.
-    pub start_line: u64,
-
-    /// The zero-based character offset from where the folded range starts. If not defined, defaults to the length of the start line.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub start_character: Option<u64>,
-
-    /// The zero-based line number where the folded range ends.
-    pub end_line: u64,
-
-    /// The zero-based character offset before the folded range ends. If not defined, defaults to the length of the end line.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub end_character: Option<u64>,
-
-    /// Describes the kind of the folding range such as `comment' or 'region'. The kind
-    /// is used to categorize folding ranges and used by commands like 'Fold all comments'. See
-    /// [FoldingRangeKind](#FoldingRangeKind) for an enumeration of standardized kinds.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub kind: Option<FoldingRangeKind>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-pub struct SelectionRangeOptions {
-    #[serde(flatten)]
-    pub work_done_progress_options: WorkDoneProgressOptions,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-pub struct SelectionRangeRegistrationOptions {
-    #[serde(flatten)]
-    pub selection_range_options: SelectionRangeOptions,
-
-    #[serde(flatten)]
-    pub registration_options: StaticTextDocumentRegistrationOptions,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(untagged)]
-pub enum SelectionRangeProviderCapability {
-    Simple(bool),
-    Options(SelectionRangeOptions),
-    RegistrationOptions(SelectionRangeRegistrationOptions),
-}
-
-impl From<SelectionRangeRegistrationOptions> for SelectionRangeProviderCapability {
-    fn from(from: SelectionRangeRegistrationOptions) -> Self {
-        Self::RegistrationOptions(from)
-    }
-}
-
-impl From<SelectionRangeOptions> for SelectionRangeProviderCapability {
-    fn from(from: SelectionRangeOptions) -> Self {
-        Self::Options(from)
-    }
-}
-
-impl From<bool> for SelectionRangeProviderCapability {
-    fn from(from: bool) -> Self {
-        Self::Simple(from)
-    }
-}
-
-/// A parameter literal used in selection range requests.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SelectionRangeParams {
-    /// The text document.
-    pub text_document: TextDocumentIdentifier,
-
-    /// The positions inside the text document.
-    pub positions: Vec<Position>,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-}
-
-/// Represents a selection range.
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SelectionRange {
-    /// Range of the selection.
-    pub range: Range,
-
-    /// The parent selection range containing this range.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub parent: Option<Box<SelectionRange>>,
-}
-
-/// Enum of known range kinds
-#[derive(Debug, Eq, PartialEq, Deserialize, Serialize, Clone)]
-#[serde(rename_all = "lowercase")]
-pub enum FoldingRangeKind {
-    /// Folding range for a comment
-    Comment,
-    /// Folding range for a imports or includes
-    Imports,
-    /// Folding range for a region (e.g. `#region`)
-    Region,
 }
 
 /// Describes the content type that a client supports in various
@@ -3592,770 +1953,12 @@ pub struct MarkupContent {
     pub value: String,
 }
 
-pub type ProgressToken = NumberOrString;
-
-/// The progress notification is sent from the server to the client to ask
-/// the client to indicate progress.
-#[derive(Debug, PartialEq, Deserialize, Serialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct ProgressParams {
-    /// The progress token provided by the client.
-    pub token: ProgressToken,
-
-    /// The progress data.
-    pub value: ProgressParamsValue,
-}
-
-#[derive(Debug, PartialEq, Deserialize, Serialize, Clone)]
-#[serde(untagged)]
-pub enum ProgressParamsValue {
-    WorkDone(WorkDoneProgress),
-}
-
-/// The `window/workDoneProgress/create` request is sent from the server
-/// to the clientto ask the client to create a work done progress.
-#[derive(Debug, PartialEq, Deserialize, Serialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkDoneProgressCreateParams {
-    /// The token to be used to report progress.
-    pub token: ProgressToken,
-}
-
-/// The `window/workDoneProgress/cancel` notification is sent from the client
-/// to the server to cancel a progress initiated on the server side using the `window/workDoneProgress/create`.
-#[derive(Debug, PartialEq, Deserialize, Serialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkDoneProgressCancelParams {
-    /// The token to be used to report progress.
-    pub token: ProgressToken,
-}
-
-/// Options to signal work done progress support in server capabilities.
-#[derive(Debug, Eq, PartialEq, Default, Deserialize, Serialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkDoneProgressOptions {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub work_done_progress: Option<bool>,
-}
-
-/// An optional token that a server can use to report work done progress
-#[derive(Debug, Eq, PartialEq, Default, Deserialize, Serialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkDoneProgressParams {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub work_done_token: Option<ProgressToken>,
-}
-
-#[derive(Debug, PartialEq, Default, Deserialize, Serialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkDoneProgressBegin {
-    /// Mandatory title of the progress operation. Used to briefly inform
-    /// about the kind of operation being performed.
-    /// Examples: "Indexing" or "Linking dependencies".
-    pub title: String,
-
-    /// Controls if a cancel button should show to allow the user to cancel the
-    /// long running operation. Clients that don't support cancellation are allowed
-    /// to ignore the setting.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cancellable: Option<bool>,
-
-    /// Optional, more detailed associated progress message. Contains
-    /// complementary information to the `title`.
-    /// Examples: "3/25 files", "project/src/module2", "node_modules/some_dep".
-    /// If unset, the previous progress message (if any) is still valid.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub message: Option<String>,
-
-    /// Optional progress percentage to display (value 100 is considered 100%).
-    /// If unset, the previous progress percentage (if any) is still valid.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub percentage: Option<f64>,
-}
-
-#[derive(Debug, PartialEq, Default, Deserialize, Serialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkDoneProgressReport {
-    /// Controls if a cancel button should show to allow the user to cancel the
-    /// long running operation. Clients that don't support cancellation are allowed
-    /// to ignore the setting.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cancellable: Option<bool>,
-
-    /// Optional, more detailed associated progress message. Contains
-    /// complementary information to the `title`.
-    /// Examples: "3/25 files", "project/src/module2", "node_modules/some_dep".
-    /// If unset, the previous progress message (if any) is still valid.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub message: Option<String>,
-
-    /// Optional progress percentage to display (value 100 is considered 100%).
-    /// If unset, the previous progress percentage (if any) is still valid.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub percentage: Option<f64>,
-}
-
-#[derive(Debug, PartialEq, Default, Deserialize, Serialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkDoneProgressEnd {
-    /// Optional, more detailed associated progress message. Contains
-    /// complementary information to the `title`.
-    /// Examples: "3/25 files", "project/src/module2", "node_modules/some_dep".
-    /// If unset, the previous progress message (if any) is still valid.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub message: Option<String>,
-}
-
-#[derive(Debug, PartialEq, Deserialize, Serialize, Clone)]
-#[serde(tag = "kind", rename_all = "lowercase")]
-pub enum WorkDoneProgress {
-    Begin(WorkDoneProgressBegin),
-    Report(WorkDoneProgressReport),
-    End(WorkDoneProgressEnd),
-}
-
 /// A parameter literal used to pass a partial result token.
 #[derive(Debug, Eq, PartialEq, Default, Deserialize, Serialize, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct PartialResultParams {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub partial_result_token: Option<ProgressToken>,
-}
-
-#[derive(Debug, Eq, PartialEq, Default, Deserialize, Serialize, Clone)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticHighlightingClientCapability {
-    /// `true` if the client supports semantic highlighting support text documents. Otherwise, `false`. It is `false` by default.
-    pub semantic_highlighting: bool,
-}
-
-#[derive(Debug, Eq, PartialEq, Default, Deserialize, Serialize, Clone)]
-#[cfg(feature = "proposed")]
-pub struct SemanticHighlightingServerCapability {
-    /// A "lookup table" of semantic highlighting [TextMate scopes](https://manual.macromates.com/en/language_grammars)
-    /// supported by the language server. If not defined or empty, then the server does not support the semantic highlighting
-    /// feature. Otherwise, clients should reuse this "lookup table" when receiving semantic highlighting notifications from
-    /// the server.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub scopes: Option<Vec<Vec<String>>>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone)]
-#[cfg(feature = "proposed")]
-pub struct SemanticHighlightingToken {
-    pub character: u32,
-    pub length: u16,
-    pub scope: u16,
-}
-
-#[cfg(feature = "proposed")]
-impl SemanticHighlightingToken {
-    /// Deserializes the tokens from a base64 encoded string
-    fn deserialize_tokens<'de, D>(
-        deserializer: D,
-    ) -> Result<Option<Vec<SemanticHighlightingToken>>, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let opt_s = Option::<String>::deserialize(deserializer)?;
-
-        if let Some(s) = opt_s {
-            let bytes = base64::decode_config(s.as_str(), base64::STANDARD)
-                .map_err(|_| serde::de::Error::custom("Error parsing base64 string"))?;
-            let mut res = Vec::new();
-            for chunk in bytes.chunks_exact(8) {
-                res.push(SemanticHighlightingToken {
-                    character: u32::from_be_bytes(<[u8; 4]>::try_from(&chunk[0..4]).unwrap()),
-                    length: u16::from_be_bytes(<[u8; 2]>::try_from(&chunk[4..6]).unwrap()),
-                    scope: u16::from_be_bytes(<[u8; 2]>::try_from(&chunk[6..8]).unwrap()),
-                });
-            }
-            Result::Ok(Some(res))
-        } else {
-            Result::Ok(None)
-        }
-    }
-
-    /// Serialize the tokens to a base64 encoded string
-    fn serialize_tokens<S>(
-        tokens: &Option<Vec<SemanticHighlightingToken>>,
-        serializer: S,
-    ) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        if let Some(tokens) = tokens {
-            let mut bytes = vec![];
-            for token in tokens {
-                bytes.extend_from_slice(&token.character.to_be_bytes());
-                bytes.extend_from_slice(&token.length.to_be_bytes());
-                bytes.extend_from_slice(&token.scope.to_be_bytes());
-            }
-            serializer.collect_str(&base64::display::Base64Display::with_config(
-                &bytes,
-                base64::STANDARD,
-            ))
-        } else {
-            serializer.serialize_none()
-        }
-    }
-}
-
-/// Represents a semantic highlighting information that has to be applied on a specific line of the text document.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[cfg(feature = "proposed")]
-pub struct SemanticHighlightingInformation {
-    /// The zero-based line position in the text document.
-    pub line: i32,
-
-    /// A base64 encoded string representing every single highlighted characters with its start position, length and the "lookup table" index of
-    /// of the semantic highlighting [TextMate scopes](https://manual.macromates.com/en/language_grammars).
-    /// If the `tokens` is empty or not defined, then no highlighted positions are available for the line.
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        deserialize_with = "SemanticHighlightingToken::deserialize_tokens",
-        serialize_with = "SemanticHighlightingToken::serialize_tokens"
-    )]
-    pub tokens: Option<Vec<SemanticHighlightingToken>>,
-}
-
-/// Parameters for the semantic highlighting (server-side) push notification.
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticHighlightingParams {
-    /// The text document that has to be decorated with the semantic highlighting information.
-    pub text_document: VersionedTextDocumentIdentifier,
-
-    /// An array of semantic highlighting information.
-    pub lines: Vec<SemanticHighlightingInformation>,
-}
-
-/// A set of predefined token types. This set is not fixed
-/// and clients can specify additional token types via the
-/// corresponding client capabilities.
-///
-/// @since 3.16.0 - Proposed state
-#[derive(Debug, Eq, PartialEq, Hash, PartialOrd, Clone, Deserialize, Serialize)]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokenType(Cow<'static, str>);
-
-#[cfg(feature = "proposed")]
-impl SemanticTokenType {
-    pub const NAMESPACE: SemanticTokenType = SemanticTokenType::new("namespace");
-    pub const TYPE: SemanticTokenType = SemanticTokenType::new("type");
-    pub const CLASS: SemanticTokenType = SemanticTokenType::new("class");
-    pub const ENUM: SemanticTokenType = SemanticTokenType::new("enum");
-    pub const INTERFACE: SemanticTokenType = SemanticTokenType::new("interface");
-    pub const STRUCT: SemanticTokenType = SemanticTokenType::new("struct");
-    pub const TYPE_PARAMETER: SemanticTokenType = SemanticTokenType::new("typeParameter");
-    pub const PARAMETER: SemanticTokenType = SemanticTokenType::new("parameter");
-    pub const VARIABLE: SemanticTokenType = SemanticTokenType::new("variable");
-    pub const PROPERTY: SemanticTokenType = SemanticTokenType::new("property");
-    pub const ENUM_MEMBER: SemanticTokenType = SemanticTokenType::new("enumMember");
-    pub const EVENT: SemanticTokenType = SemanticTokenType::new("event");
-    pub const FUNCTION: SemanticTokenType = SemanticTokenType::new("function");
-    pub const MEMBER: SemanticTokenType = SemanticTokenType::new("member");
-    pub const MACRO: SemanticTokenType = SemanticTokenType::new("macro");
-    pub const KEYWORD: SemanticTokenType = SemanticTokenType::new("keyword");
-    pub const MODIFIER: SemanticTokenType = SemanticTokenType::new("modifier");
-    pub const COMMENT: SemanticTokenType = SemanticTokenType::new("comment");
-    pub const STRING: SemanticTokenType = SemanticTokenType::new("string");
-    pub const NUMBER: SemanticTokenType = SemanticTokenType::new("number");
-    pub const REGEXP: SemanticTokenType = SemanticTokenType::new("regexp");
-    pub const OPERATOR: SemanticTokenType = SemanticTokenType::new("operator");
-
-    pub const fn new(tag: &'static str) -> Self {
-        SemanticTokenType(Cow::Borrowed(tag))
-    }
-
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-#[cfg(feature = "proposed")]
-impl From<String> for SemanticTokenType {
-    fn from(from: String) -> Self {
-        SemanticTokenType(Cow::from(from))
-    }
-}
-#[cfg(feature = "proposed")]
-impl From<&'static str> for SemanticTokenType {
-    fn from(from: &'static str) -> Self {
-        SemanticTokenType::new(from)
-    }
-}
-
-/// A set of predefined token modifiers. This set is not fixed
-/// and clients can specify additional token types via the
-/// corresponding client capabilities.
-///
-/// @since 3.16.0 - Proposed state
-#[derive(Debug, Eq, PartialEq, Hash, PartialOrd, Clone, Deserialize, Serialize)]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokenModifier(Cow<'static, str>);
-
-#[cfg(feature = "proposed")]
-impl SemanticTokenModifier {
-    pub const DECLARATION: SemanticTokenModifier = SemanticTokenModifier::new("declaration");
-    pub const DEFINITION: SemanticTokenModifier = SemanticTokenModifier::new("definition");
-    pub const READONLY: SemanticTokenModifier = SemanticTokenModifier::new("readonly");
-    pub const STATIC: SemanticTokenModifier = SemanticTokenModifier::new("static");
-    pub const DEPRECATED: SemanticTokenModifier = SemanticTokenModifier::new("deprecated");
-    pub const ABSTRACT: SemanticTokenModifier = SemanticTokenModifier::new("abstract");
-    pub const ASYNC: SemanticTokenModifier = SemanticTokenModifier::new("async");
-    pub const MODIFICATION: SemanticTokenModifier = SemanticTokenModifier::new("modification");
-    pub const DOCUMENTATION: SemanticTokenModifier = SemanticTokenModifier::new("documentation");
-    pub const DEFAULT_LIBRARY: SemanticTokenModifier = SemanticTokenModifier::new("defaultLibrary");
-
-    pub const fn new(tag: &'static str) -> Self {
-        SemanticTokenModifier(Cow::Borrowed(tag))
-    }
-
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-#[cfg(feature = "proposed")]
-impl From<String> for SemanticTokenModifier {
-    fn from(from: String) -> Self {
-        SemanticTokenModifier(Cow::from(from))
-    }
-}
-#[cfg(feature = "proposed")]
-impl From<&'static str> for SemanticTokenModifier {
-    fn from(from: &'static str) -> Self {
-        SemanticTokenModifier::new(from)
-    }
-}
-
-#[derive(Debug, Eq, PartialEq, Hash, PartialOrd, Clone, Deserialize, Serialize)]
-#[cfg(feature = "proposed")]
-pub struct TokenFormat(Cow<'static, str>);
-
-#[cfg(feature = "proposed")]
-impl TokenFormat {
-    pub const RELATIVE: TokenFormat = TokenFormat::new("relative");
-
-    pub const fn new(tag: &'static str) -> Self {
-        TokenFormat(Cow::Borrowed(tag))
-    }
-
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-#[cfg(feature = "proposed")]
-impl From<String> for TokenFormat {
-    fn from(from: String) -> Self {
-        TokenFormat(Cow::from(from))
-    }
-}
-#[cfg(feature = "proposed")]
-impl From<&'static str> for TokenFormat {
-    fn from(from: &'static str) -> Self {
-        TokenFormat::new(from)
-    }
-}
-
-/// @since 3.16.0 - Proposed state
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokensLegend {
-    /// The token types a server uses.
-    pub token_types: Vec<SemanticTokenType>,
-
-    /// The token modifiers a server uses.
-    pub token_modifiers: Vec<SemanticTokenModifier>,
-}
-
-/// The actual tokens. For a detailed description about how the data is
-/// structured please see
-/// https://github.com/microsoft/vscode-extension-samples/blob/5ae1f7787122812dcc84e37427ca90af5ee09f14/semantic-tokens-sample/vscode.proposed.d.ts#L71
-#[derive(Debug, Eq, PartialEq, Copy, Clone, Default)]
-#[cfg(feature = "proposed")]
-pub struct SemanticToken {
-    pub delta_line: u32,
-    pub delta_start: u32,
-    pub length: u32,
-    pub token_type: u32,
-    pub token_modifiers_bitset: u32,
-}
-
-#[cfg(feature = "proposed")]
-impl SemanticToken {
-    fn deserialize_tokens<'de, D>(deserializer: D) -> Result<Vec<SemanticToken>, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let data = Vec::<u32>::deserialize(deserializer)?;
-        let chunks = data.chunks_exact(5);
-
-        if !chunks.remainder().is_empty() {
-            return Result::Err(serde::de::Error::custom("Length is not divisible by 5"));
-        }
-
-        Result::Ok(
-            chunks
-                .map(|chunk| SemanticToken {
-                    delta_line: chunk[0],
-                    delta_start: chunk[1],
-                    length: chunk[2],
-                    token_type: chunk[3],
-                    token_modifiers_bitset: chunk[4],
-                })
-                .collect(),
-        )
-    }
-
-    fn serialize_tokens<S>(tokens: &[SemanticToken], serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        let mut seq = serializer.serialize_seq(Some(tokens.len() * 5))?;
-        for token in tokens.iter() {
-            seq.serialize_element(&token.delta_line)?;
-            seq.serialize_element(&token.delta_start)?;
-            seq.serialize_element(&token.length)?;
-            seq.serialize_element(&token.token_type)?;
-            seq.serialize_element(&token.token_modifiers_bitset)?;
-        }
-        seq.end()
-    }
-
-    fn deserialize_tokens_opt<'de, D>(
-        deserializer: D,
-    ) -> Result<Option<Vec<SemanticToken>>, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        #[derive(Deserialize)]
-        #[serde(transparent)]
-        struct Wrapper {
-            #[serde(deserialize_with = "SemanticToken::deserialize_tokens")]
-            tokens: Vec<SemanticToken>,
-        }
-
-        Ok(Option::<Wrapper>::deserialize(deserializer)?.map(|wrapper| wrapper.tokens))
-    }
-
-    fn serialize_tokens_opt<S>(
-        data: &Option<Vec<SemanticToken>>,
-        serializer: S,
-    ) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        #[derive(Serialize)]
-        #[serde(transparent)]
-        struct Wrapper {
-            #[serde(serialize_with = "SemanticToken::serialize_tokens")]
-            tokens: Vec<SemanticToken>,
-        }
-
-        let opt = data.as_ref().map(|t| Wrapper { tokens: t.to_vec() });
-
-        opt.serialize(serializer)
-    }
-}
-
-/// @since 3.16.0 - Proposed state
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokens {
-    /// An optional result id. If provided and clients support delta updating
-    /// the client will include the result id in the next semantic token request.
-    /// A server can then instead of computing all semantic tokens again simply
-    /// send a delta.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub result_id: Option<String>,
-
-    /// The actual tokens. For a detailed description about how the data is
-    /// structured please see
-    /// https://github.com/microsoft/vscode-extension-samples/blob/5ae1f7787122812dcc84e37427ca90af5ee09f14/semantic-tokens-sample/vscode.proposed.d.ts#L71
-    #[serde(
-        deserialize_with = "SemanticToken::deserialize_tokens",
-        serialize_with = "SemanticToken::serialize_tokens"
-    )]
-    pub data: Vec<SemanticToken>,
-}
-
-/// @since 3.16.0 - Proposed state
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokensPartialResult {
-    #[serde(
-        deserialize_with = "SemanticToken::deserialize_tokens",
-        serialize_with = "SemanticToken::serialize_tokens"
-    )]
-    pub data: Vec<SemanticToken>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[serde(untagged)]
-#[cfg(feature = "proposed")]
-pub enum SemanticTokensResult {
-    Tokens(SemanticTokens),
-    Partial(SemanticTokensPartialResult),
-}
-
-#[cfg(feature = "proposed")]
-impl From<SemanticTokens> for SemanticTokensResult {
-    fn from(from: SemanticTokens) -> Self {
-        SemanticTokensResult::Tokens(from)
-    }
-}
-
-#[cfg(feature = "proposed")]
-impl From<SemanticTokensPartialResult> for SemanticTokensResult {
-    fn from(from: SemanticTokensPartialResult) -> Self {
-        SemanticTokensResult::Partial(from)
-    }
-}
-
-/// @since 3.16.0 - Proposed state
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokensEdit {
-    pub start: u32,
-    pub delete_count: u32,
-
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        deserialize_with = "SemanticToken::deserialize_tokens_opt",
-        serialize_with = "SemanticToken::serialize_tokens_opt"
-    )]
-    pub data: Option<Vec<SemanticToken>>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[serde(untagged)]
-#[cfg(feature = "proposed")]
-pub enum SemanticTokensFullDeltaResult {
-    Tokens(SemanticTokens),
-    TokensDelta(SemanticTokensDelta),
-    PartialTokensDelta { edits: Vec<SemanticTokensEdit> },
-}
-
-#[cfg(feature = "proposed")]
-impl From<SemanticTokens> for SemanticTokensFullDeltaResult {
-    fn from(from: SemanticTokens) -> Self {
-        SemanticTokensFullDeltaResult::Tokens(from)
-    }
-}
-
-#[cfg(feature = "proposed")]
-impl From<SemanticTokensDelta> for SemanticTokensFullDeltaResult {
-    fn from(from: SemanticTokensDelta) -> Self {
-        SemanticTokensFullDeltaResult::TokensDelta(from)
-    }
-}
-
-/// @since 3.16.0 - Proposed state
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokensDelta {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub result_id: Option<String>,
-    /// For a detailed description how these edits are structured pls see
-    /// https://github.com/microsoft/vscode-extension-samples/blob/5ae1f7787122812dcc84e37427ca90af5ee09f14/semantic-tokens-sample/vscode.proposed.d.ts#L131
-    pub edits: Vec<SemanticTokensEdit>,
-}
-
-/// Capabilities specific to the `textDocument/semanticTokens/*` requests.
-///
-/// @since 3.16.0 - Proposed state
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokensClientCapabilities {
-    /// Whether implementation supports dynamic registration. If this is set to `true`
-    /// the client supports the new `(TextDocumentRegistrationOptions & StaticRegistrationOptions)`
-    /// return value for the corresponding server capability as well.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub dynamic_registration: Option<bool>,
-
-    /// Which requests the client supports and might send to the server.
-    pub requests: SemanticTokensClientCapabilitiesRequests,
-
-    /// The token types that the client supports.
-    pub token_types: Vec<SemanticTokenType>,
-
-    /// The token modifiers that the client supports.
-    pub token_modifiers: Vec<SemanticTokenModifier>,
-
-    /// The formats the clients supports.
-    pub formats: Vec<TokenFormat>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokensClientCapabilitiesRequests {
-    /// The client will send the `textDocument/semanticTokens/range` request if the server provides a corresponding handler.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub range: Option<bool>,
-
-    /// The client will send the `textDocument/semanticTokens/full` request if the server provides a corresponding handler.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub full: Option<SemanticTokensFullOptions>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[serde(untagged)]
-#[cfg(feature = "proposed")]
-pub enum SemanticTokensFullOptions {
-    Bool(bool),
-    Delta {
-        /// The client will send the `textDocument/semanticTokens/full/delta` request if the server provides a corresponding handler.
-        /// The server supports deltas for full documents.
-        #[serde(skip_serializing_if = "Option::is_none")]
-        delta: Option<bool>,
-    },
-}
-
-/// @since 3.16.0 - Proposed state
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokensOptions {
-    #[serde(flatten)]
-    pub work_done_progress_options: WorkDoneProgressOptions,
-
-    /// The legend used by the server
-    pub legend: SemanticTokensLegend,
-
-    /// Server supports providing semantic tokens for a sepcific range
-    /// of a document.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub range: Option<bool>,
-
-    /// Server supports providing semantic tokens for a full document.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub full: Option<SemanticTokensFullOptions>,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokensRegistrationOptions {
-    #[serde(flatten)]
-    pub text_document_registration_options: TextDocumentRegistrationOptions,
-
-    #[serde(flatten)]
-    pub semantic_tokens_options: SemanticTokensOptions,
-
-    #[serde(flatten)]
-    pub static_registration_options: StaticRegistrationOptions,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[serde(untagged)]
-#[cfg(feature = "proposed")]
-pub enum SemanticTokensServerCapabilities {
-    SemanticTokensOptions(SemanticTokensOptions),
-    SemanticTokensRegistrationOptions(SemanticTokensRegistrationOptions),
-}
-
-#[cfg(feature = "proposed")]
-impl From<SemanticTokensOptions> for SemanticTokensServerCapabilities {
-    fn from(from: SemanticTokensOptions) -> Self {
-        SemanticTokensServerCapabilities::SemanticTokensOptions(from)
-    }
-}
-
-#[cfg(feature = "proposed")]
-impl From<SemanticTokensRegistrationOptions> for SemanticTokensServerCapabilities {
-    fn from(from: SemanticTokensRegistrationOptions) -> Self {
-        SemanticTokensServerCapabilities::SemanticTokensRegistrationOptions(from)
-    }
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokensParams {
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-
-    /// The text document.
-    pub text_document: TextDocumentIdentifier,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokensDeltaParams {
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-
-    /// The text document.
-    pub text_document: TextDocumentIdentifier,
-
-    /// The result id of a previous response. The result Id can either point to a full response
-    /// or a delta response depending on what was recevied last.
-    pub previous_result_id: String,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct SemanticTokensRangeParams {
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-
-    /// The text document.
-    pub text_document: TextDocumentIdentifier,
-
-    /// The range the semantic tokens are requested for.
-    pub range: Range,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[serde(untagged)]
-#[cfg(feature = "proposed")]
-pub enum SemanticTokensRangeResult {
-    Tokens(SemanticTokens),
-    Partial(SemanticTokensPartialResult),
-}
-
-#[cfg(feature = "proposed")]
-impl From<SemanticTokens> for SemanticTokensRangeResult {
-    fn from(tokens: SemanticTokens) -> Self {
-        SemanticTokensRangeResult::Tokens(tokens)
-    }
-}
-
-#[cfg(feature = "proposed")]
-impl From<SemanticTokensPartialResult> for SemanticTokensRangeResult {
-    fn from(partial: SemanticTokensPartialResult) -> Self {
-        SemanticTokensRangeResult::Partial(partial)
-    }
 }
 
 /// Symbol tags are extra annotations that tweak the rendering of a symbol.
@@ -4368,135 +1971,12 @@ pub enum SymbolTag {
     Deprecated = 1,
 }
 
-#[derive(Debug, Eq, PartialEq, Clone, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct CallHierarchyOptions {
-    #[serde(flatten)]
-    pub work_done_progress_options: WorkDoneProgressOptions,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(untagged)]
-#[cfg(feature = "proposed")]
-pub enum CallHierarchyServerCapability {
-    Simple(bool),
-    Options(CallHierarchyOptions),
-}
-
-#[cfg(feature = "proposed")]
-impl From<CallHierarchyOptions> for CallHierarchyServerCapability {
-    fn from(from: CallHierarchyOptions) -> Self {
-        Self::Options(from)
-    }
-}
-
-#[cfg(feature = "proposed")]
-impl From<bool> for CallHierarchyServerCapability {
-    fn from(from: bool) -> Self {
-        Self::Simple(from)
-    }
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct CallHierarchyPrepareParams {
-    #[serde(flatten)]
-    pub text_document_position_params: TextDocumentPositionParams,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct CallHierarchyItem {
-    /// The name of this item.
-    pub name: String,
-
-    /// The kind of this item.
-    pub kind: SymbolKind,
-
-    /// Tags for this item.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<Vec<SymbolTag>>,
-
-    /// More detail for this item, e.g. the signature of a function.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub detail: Option<String>,
-
-    /// The resource identifier of this item.
-    pub uri: Url,
-
-    /// The range enclosing this symbol not including leading/trailing whitespace but everything else, e.g. comments and code.
-    pub range: Range,
-
-    /// The range that should be selected and revealed when this symbol is being picked, e.g. the name of a function.
-    /// Must be contained by the [`range`](#CallHierarchyItem.range).
-    pub selection_range: Range,
-}
-
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct CallHierarchyIncomingCallsParams {
-    pub item: CallHierarchyItem,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-}
-
-/// Represents an incoming call, e.g. a caller of a method or constructor.
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct CallHierarchyIncomingCall {
-    /// The item that makes the call.
-    pub from: CallHierarchyItem,
-
-    /// The range at which at which the calls appears. This is relative to the caller
-    /// denoted by [`this.from`](#CallHierarchyIncomingCall.from).
-    pub from_ranges: Vec<Range>,
-}
-
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct CallHierarchyOutgoingCallsParams {
-    pub item: CallHierarchyItem,
-
-    #[serde(flatten)]
-    pub work_done_progress_params: WorkDoneProgressParams,
-
-    #[serde(flatten)]
-    pub partial_result_params: PartialResultParams,
-}
-
-/// Represents an outgoing call, e.g. calling a getter from a method or a method from a constructor etc.
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
-#[serde(rename_all = "camelCase")]
-#[cfg(feature = "proposed")]
-pub struct CallHierarchyOutgoingCall {
-    /// The item that is called.
-    pub to: CallHierarchyItem,
-
-    /// The range at which this item is called. This is the range relative to the caller, e.g the item
-    /// passed to [`provideCallHierarchyOutgoingCalls`](#CallHierarchyItemProvider.provideCallHierarchyOutgoingCalls)
-    /// and not [`this.to`](#CallHierarchyOutgoingCall.to).
-    pub from_ranges: Vec<Range>,
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use serde::{Deserialize, Serialize};
 
-    fn test_serialization<SER>(ms: &SER, expected: &str)
+    pub(crate) fn test_serialization<SER>(ms: &SER, expected: &str)
     where
         SER: Serialize + for<'de> Deserialize<'de> + PartialEq + std::fmt::Debug,
     {
@@ -4506,7 +1986,7 @@ mod tests {
         assert_eq!(&deserialized, ms);
     }
 
-    fn test_deserialization<T>(json: &str, expected: &T)
+    pub(crate) fn test_deserialization<T>(json: &str, expected: &T)
     where
         T: for<'de> Deserialize<'de> + PartialEq + std::fmt::Debug,
     {
@@ -4574,35 +2054,6 @@ mod tests {
     }
 
     #[test]
-    fn formatting_options() {
-        test_serialization(
-            &FormattingOptions {
-                tab_size: 123,
-                insert_spaces: true,
-                properties: HashMap::new(),
-                trim_trailing_whitespace: None,
-                insert_final_newline: None,
-                trim_final_newlines: None,
-            },
-            r#"{"tabSize":123,"insertSpaces":true}"#,
-        );
-
-        test_serialization(
-            &FormattingOptions {
-                tab_size: 123,
-                insert_spaces: true,
-                properties: vec![("prop".to_string(), FormattingProperty::Number(1.0))]
-                    .into_iter()
-                    .collect(),
-                trim_trailing_whitespace: None,
-                insert_final_newline: None,
-                trim_final_newlines: None,
-            },
-            r#"{"tabSize":123,"insertSpaces":true,"prop":1.0}"#,
-        );
-    }
-
-    #[test]
     fn root_uri_can_be_missing() {
         serde_json::from_str::<InitializeParams>(r#"{ "capabilities": {} }"#).unwrap();
     }
@@ -4626,263 +2077,6 @@ mod tests {
                 ResourceOperationKind::Delete,
             ],
             r#"["create","rename","delete"]"#,
-        );
-    }
-
-    #[test]
-    fn test_code_action_response() {
-        test_serialization(
-            &vec![
-                CodeActionOrCommand::Command(Command {
-                    title: "title".to_string(),
-                    command: "command".to_string(),
-                    arguments: None,
-                }),
-                CodeActionOrCommand::CodeAction(CodeAction {
-                    title: "title".to_string(),
-                    kind: Some(CodeActionKind::QUICKFIX),
-                    command: None,
-                    diagnostics: None,
-                    edit: None,
-                    is_preferred: None,
-                }),
-            ],
-            r#"[{"title":"title","command":"command"},{"title":"title","kind":"quickfix"}]"#,
-        )
-    }
-
-    #[cfg(feature = "proposed")]
-    #[test]
-    fn test_semantic_highlighting_information_serialization() {
-        test_serialization(
-            &SemanticHighlightingInformation {
-                line: 10,
-                tokens: Some(vec![
-                    SemanticHighlightingToken {
-                        character: 0x00000001,
-                        length: 0x0002,
-                        scope: 0x0003,
-                    },
-                    SemanticHighlightingToken {
-                        character: 0x00112222,
-                        length: 0x0FF0,
-                        scope: 0x0202,
-                    },
-                ]),
-            },
-            r#"{"line":10,"tokens":"AAAAAQACAAMAESIiD/ACAg=="}"#,
-        );
-
-        test_serialization(
-            &SemanticHighlightingInformation {
-                line: 22,
-                tokens: None,
-            },
-            r#"{"line":22}"#,
-        );
-    }
-
-    #[test]
-    fn test_tag_support_deserialization() {
-        let mut empty = CompletionItemCapability::default();
-        empty.tag_support = None;
-
-        test_deserialization(r#"{}"#, &empty);
-        test_deserialization(r#"{"tagSupport": false}"#, &empty);
-
-        let mut t = CompletionItemCapability::default();
-        t.tag_support = Some(TagSupport { value_set: vec![] });
-        test_deserialization(r#"{"tagSupport": true}"#, &t);
-
-        let mut t = CompletionItemCapability::default();
-        t.tag_support = Some(TagSupport {
-            value_set: vec![CompletionItemTag::Deprecated],
-        });
-        test_deserialization(r#"{"tagSupport": {"valueSet": [1]}}"#, &t);
-    }
-
-    #[cfg(feature = "proposed")]
-    #[test]
-    fn test_semantic_tokens_support_serialization() {
-        test_serialization(
-            &SemanticTokens {
-                result_id: None,
-                data: vec![],
-            },
-            r#"{"data":[]}"#,
-        );
-
-        test_serialization(
-            &SemanticTokens {
-                result_id: None,
-                data: vec![SemanticToken {
-                    delta_line: 2,
-                    delta_start: 5,
-                    length: 3,
-                    token_type: 0,
-                    token_modifiers_bitset: 3,
-                }],
-            },
-            r#"{"data":[2,5,3,0,3]}"#,
-        );
-
-        test_serialization(
-            &SemanticTokens {
-                result_id: None,
-                data: vec![
-                    SemanticToken {
-                        delta_line: 2,
-                        delta_start: 5,
-                        length: 3,
-                        token_type: 0,
-                        token_modifiers_bitset: 3,
-                    },
-                    SemanticToken {
-                        delta_line: 0,
-                        delta_start: 5,
-                        length: 4,
-                        token_type: 1,
-                        token_modifiers_bitset: 0,
-                    },
-                ],
-            },
-            r#"{"data":[2,5,3,0,3,0,5,4,1,0]}"#,
-        );
-    }
-
-    #[cfg(feature = "proposed")]
-    #[test]
-    fn test_semantic_tokens_support_deserialization() {
-        test_deserialization(
-            r#"{"data":[]}"#,
-            &SemanticTokens {
-                result_id: None,
-                data: vec![],
-            },
-        );
-
-        test_deserialization(
-            r#"{"data":[2,5,3,0,3]}"#,
-            &SemanticTokens {
-                result_id: None,
-                data: vec![SemanticToken {
-                    delta_line: 2,
-                    delta_start: 5,
-                    length: 3,
-                    token_type: 0,
-                    token_modifiers_bitset: 3,
-                }],
-            },
-        );
-
-        test_deserialization(
-            r#"{"data":[2,5,3,0,3,0,5,4,1,0]}"#,
-            &SemanticTokens {
-                result_id: None,
-                data: vec![
-                    SemanticToken {
-                        delta_line: 2,
-                        delta_start: 5,
-                        length: 3,
-                        token_type: 0,
-                        token_modifiers_bitset: 3,
-                    },
-                    SemanticToken {
-                        delta_line: 0,
-                        delta_start: 5,
-                        length: 4,
-                        token_type: 1,
-                        token_modifiers_bitset: 0,
-                    },
-                ],
-            },
-        );
-    }
-
-    #[cfg(feature = "proposed")]
-    #[test]
-    #[should_panic]
-    fn test_semantic_tokens_support_deserialization_err() {
-        test_deserialization(
-            r#"{"data":[1]}"#,
-            &SemanticTokens {
-                result_id: None,
-                data: vec![],
-            },
-        );
-    }
-
-    #[cfg(feature = "proposed")]
-    #[test]
-    fn test_semantic_tokens_edit_support_deserialization() {
-        test_deserialization(
-            r#"{"start":0,"deleteCount":1,"data":[2,5,3,0,3,0,5,4,1,0]}"#,
-            &SemanticTokensEdit {
-                start: 0,
-                delete_count: 1,
-                data: Some(vec![
-                    SemanticToken {
-                        delta_line: 2,
-                        delta_start: 5,
-                        length: 3,
-                        token_type: 0,
-                        token_modifiers_bitset: 3,
-                    },
-                    SemanticToken {
-                        delta_line: 0,
-                        delta_start: 5,
-                        length: 4,
-                        token_type: 1,
-                        token_modifiers_bitset: 0,
-                    },
-                ]),
-            },
-        );
-
-        test_deserialization(
-            r#"{"start":0,"deleteCount":1}"#,
-            &SemanticTokensEdit {
-                start: 0,
-                delete_count: 1,
-                data: None,
-            },
-        );
-    }
-
-    #[cfg(feature = "proposed")]
-    #[test]
-    fn test_semantic_tokens_edit_support_serialization() {
-        test_serialization(
-            &SemanticTokensEdit {
-                start: 0,
-                delete_count: 1,
-                data: Some(vec![
-                    SemanticToken {
-                        delta_line: 2,
-                        delta_start: 5,
-                        length: 3,
-                        token_type: 0,
-                        token_modifiers_bitset: 3,
-                    },
-                    SemanticToken {
-                        delta_line: 0,
-                        delta_start: 5,
-                        length: 4,
-                        token_type: 1,
-                        token_modifiers_bitset: 0,
-                    },
-                ]),
-            },
-            r#"{"start":0,"deleteCount":1,"data":[2,5,3,0,3,0,5,4,1,0]}"#,
-        );
-
-        test_serialization(
-            &SemanticTokensEdit {
-                start: 0,
-                delete_count: 1,
-                data: None,
-            },
-            r#"{"start":0,"deleteCount":1}"#,
         );
     }
 }
